+++ src/object-enchant/object-smith.h
@@ -9,6 +9,8 @@
 
 struct object_type;
 struct player_type;
+struct smith_info_type;
+struct essence_drain_type;
 class ItemTester;
 
 enum class SmithEffect;
@@ -45,5 +47,12 @@ public:
     int get_addable_count(SmithEffect smith_effect, int item_number) const;
 
 private:
+    static std::optional<const smith_info_type *> find_smith_info(SmithEffect effect);
+
+    static const std::vector<SmithEssence> essence_list_order;
+    static const std::unordered_map<SmithEssence, concptr> essence_to_name;
+    static const std::vector<essence_drain_type> essence_drain_info_table;
+    static const std::vector<smith_info_type> smith_info_table;
+
     player_type *player_ptr;
 };
